requirements.txt
# AWS SDK
boto3>=1.28.0

# API / Validação
pydantic>=2.0.0

# Autenticação
PyJWT>=2.8.0
bcrypt>=4.0.0

# Testes
pytest>=7.4.0
pytest-cov>=4.1.0
moto>=4.2.0

# Linting
ruff>=0.1.0
